Update nextRTC.js
There was in error in setChannelReady.  The 'in' operator in javascript will iterate over members of an object (not just array elements) as opposed to how it works in java.  So my client was sending 'addAll' to the server and crashing the connection.   (Add all wasn't waiting to be sent, but was a method of the empty array).   I changed to an old school loop.